apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
aiomysql==0.2.0
pymysql==1.1.0
python-dotenv==1.0.1
httpx==0.27.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
torch==2.0.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
redis==5.0.1
